pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "goldflipper"
version = "0.2.5"
description = "A semi-autonomous options trading system on Alpaca."
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "Proprietary" }
authors = [
    { name = "Iliya Yaroshevskiy", email = "purpleafmanagement@gmail.com" },
]
dependencies = [
    "alpaca-py>=0.43.2",
    "yfinance>=1.0",
    "marketdata-sdk-py[pandas]>=1.1.0",
    "pandas>=2.3.3",
    "pandas_market_calendars>=4.6.0",
    "numpy>=2.0,<3.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "scipy>=1.16.0",
    "ta>=0.11.0",
    "textual>=6.7.0",
    "psutil>=7.1.0",
    "nest-asyncio>=1.6.0",
    "pywin32; sys_platform == 'win32'",
    "tkinterdnd2>=0.3.0",
    "requests>=2.32.0",
    "charset-normalizer>=3.4.0",
    "urllib3>=2.5.0",
    "PyYAML>=6.0.3",
    "ruamel.yaml>=0.18.6",
    "colorama>=0.4.6",
    "rich>=14.0.0",
    "mplfinance>=0.12.10b0",
    "XlsxWriter>=3.2.0",
    "fastmcp>=2.0,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.0",
    "pyright>=1.1.407",
    "nuitka>=2.0",
]

[project.scripts]
goldflipper = "goldflipper.run:main"
goldflipper-mcp = "goldflipper.mcp_server.__main__:main"

[project.urls]
Homepage = "https://github.com/Zaroganos/goldflipper"
"Bug Reports" = "https://github.com/Zaroganos/goldflipper/issues"
Documentation = "https://github.com/Zaroganos/goldflipper#readme"

[tool.hatch.build.targets.wheel]
packages = ["goldflipper"]

[tool.hatch.build.targets.wheel.force-include]
"goldflipper/tools/play-template.json" = "goldflipper/tools/play-template.json"
"goldflipper/config/settings_template.yaml" = "goldflipper/config/settings_template.yaml"
"goldflipper/reference" = "goldflipper/reference"

[tool.ruff]
line-length = 150
target-version = "py312"
src = ["goldflipper", "src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "NPY201",
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"goldflipper/chart/chart_viewer.py" = ["E402"]
"goldflipper/data/market/providers/marketdataapp_test.py" = ["E402"]
"goldflipper/run.py" = ["E402"]
"goldflipper/run_multi.py" = ["E402"]
"goldflipper/startup_test.py" = ["E402"]
"goldflipper/tools/auto_play_creator.py" = ["E402"]
"goldflipper/tools/multi_market_data.py" = ["E402"]
"goldflipper/tools/option_data_fetcher.py" = ["E402"]
"goldflipper/tools/options-data.py" = ["E402"]
"goldflipper/tools/play-csv-selftest.py" = ["E402"]
"goldflipper/tools/play_creator_gui.py" = ["E402"]
"goldflipper/tools/play_csv_ingestion_multitool.py" = ["E402"]
"goldflipper/tools/play_csv_ingestion_tool.py" = ["E402"]
"goldflipper/tools/system_status.py" = ["E402"]
"goldflipper/trade_logging/data_backfill_helper.py" = ["E402"]
"goldflipper/trade_logging/trade_logger_ui.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["goldflipper", "src"]
exclude = ["**/__pycache__", "**/.venv"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

